+++ packages/components/src/components/Align/index.ts
@@ -1,0 +1,4 @@
+import { Align } from "./Align";
+
+export { type AlignProps, Align } from "./Align";
+export default Align;