+++ packages/components/src/components/List/model/item/ItemView.ts
@@ -6,6 +6,7 @@ import type List from "@/components/List/model/List";
 export interface ItemViewShape<T> {
   textValue?: (data: T) => string;
   href?: (data: T) => string;
+  defaultExpanded?: (data: T) => boolean;
   renderFn?: RenderItemFn<T>;
   fallback?: ReactElement;
 }
@@ -14,15 +15,17 @@ export class ItemView<T> {
   public readonly list: List<T>;
   public readonly textValue?: (data: T) => string;
   public readonly href?: (data: T) => string;
+  public readonly defaultExpanded?: (data: T) => boolean;
   public readonly fallback?: ReactElement;
   private readonly renderFn?: RenderItemFn<T>;
 
   public constructor(list: List<T>, shape: ItemViewShape<T> = {}) {
-    const { fallback, textValue, href, renderFn } = shape;
+    const { fallback, textValue, href, defaultExpanded, renderFn } = shape;
     this.list = list;
     this.textValue = textValue;
     this.renderFn = renderFn;
     this.href = href;
+    this.defaultExpanded = defaultExpanded;
     this.fallback = fallback;
   }