+++ packages/components/src/components/Markdown/index.ts
@@ -1,0 +1,4 @@
+import { Markdown } from "./Markdown";
+
+export { type MarkdownProps, Markdown } from "./Markdown";
+export default Markdown;